+++ Content.Client/Entry/IgnoredComponents.cs
@@ -54,7 +54,6 @@ namespace Content.Client.Entry
             "ResearchPointSource",
             "ResearchClient",
             "IdCardConsole",
-            "MimePowers",
             "ThermalRegulator",
             "DiseaseMachineRunning",
             "DiseaseMachine",
