--- Content.Server/Abilities/Mime/MimePowersComponent.cs
@@ -1,16 +1,0 @@
-namespace Content.Server.Abilities.Mime
-{
-    /// <summary>
-    /// Lets its owner entity use mime powers, like placing invisible walls.
-    /// </summary>
-    [RegisterComponent]
-    public sealed class MimePowersComponent : Component
-    {
-        /// <summary>
-        /// Whether this component is active or not.
-        /// </summarY>
-        [ViewVariables]
-        [DataField("enabled")]
-        public bool Enabled = true;
-    }
-}
