--- Content.Server/Abilities/Mime/MimePowersSystem.cs
@@ -1,25 +1,0 @@
-using Content.Shared.Speech;
-using Content.Server.Popups;
-using Robust.Shared.Player;
-
-namespace Content.Server.Abilities.Mime
-{
-    public sealed class MimePowersSystem : EntitySystem
-    {
-        [Dependency] private readonly PopupSystem _popupSystem = default!;
-        public override void Initialize()
-        {
-            base.Initialize();
-            SubscribeLocalEvent<MimePowersComponent, SpeakAttemptEvent>(OnSpeakAttempt);
-        }
-
-        private void OnSpeakAttempt(EntityUid uid, MimePowersComponent component, SpeakAttemptEvent args)
-        {
-            if (!component.Enabled)
-                return;
-
-            _popupSystem.PopupEntity(Loc.GetString("mime-cant-speak"), uid, Filter.Entities(uid));
-            args.Cancel();
-        }
-    }
-}
