+++ Resources/Locale/en-US/abilities/mime.ftl
@@ -1,1 +1,3 @@
 mime-cant-speak = Your vow of silence prevents you from speaking.
+mime-invisible-wall = Create Invisible Wall
+mime-invisible-wall-desc = Create an invisible wall in front of you, if placeable there.
